requirements.txt
beautifulsoup4==4.12.2
playwright==1.40.0
pyyaml==6.0.1
python-docx==0.8.11
numpy==1.23.5
pandas==1.5.3
requests==2.27.1
twocaptcha==0.0.1
anticaptchaofficial==1.0.63
pydantic==1.9.0
loguru==0.7.2
PyPDF2==3.0.1
python-dotenv==1.0.0
ollama==0.1.5 